+++ src/js/view/winnerViewControl.js
@@ -1,5 +1,5 @@
-import { racingWinnerText } from "../util/elements.js";
-import { SEPARATOR } from "../util/constants.js";
+import { racingWinnerText } from "../elements.js";
+import { SEPARATOR } from "../constants.js";
 
 export const setWinnerText = winners => {
   racingWinnerText.innerHTML = winners.join(`${SEPARATOR} `);
